+++ src/components/lukso-card/index.ts
@@ -13,6 +13,7 @@ import style from './style.scss?inline'
 import '@/components/lukso-profile'
 import { cn } from '@/shared/tools'
 import { customStyleMap } from '@/shared/directives'
+import '@/components/lukso-image'
 
 export type CardVariants =
   | 'basic'
@@ -79,7 +80,9 @@ export class LuksoCard extends TailwindStyledElement(style) {
     if (this.backgroundUrl) {
       return `url('${this.backgroundUrl}')`
     }
+  }
 
+  private backgroundGradient() {
     let gradientStart = '#24354210' // bg-neutral-20/10
     let gradientEnd = '#24354220' // bg-neutral-20/20
     const opacity = '80' // 50% in hex format
@@ -93,40 +96,40 @@ export class LuksoCard extends TailwindStyledElement(style) {
   }
 
   private cardStyles = tv({
-    base: `bg-neutral-100`,
+    base: 'bg-neutral-100',
     variants: {
       borderRadius: {
-        none: `rounded-0`,
-        small: `rounded-12`,
-        medium: `rounded-24`,
+        none: 'rounded-0',
+        small: 'rounded-12',
+        medium: 'rounded-24',
       },
       shadow: {
-        small: `shadow-neutral-drop-shadow`,
-        medium: `shadow-neutral-drop-shadow-1xl`,
-        large: `shadow-neutral-drop-shadow-2xl`,
+        small: 'shadow-neutral-drop-shadow',
+        medium: 'shadow-neutral-drop-shadow-1xl',
+        large: 'shadow-neutral-drop-shadow-2xl',
       },
       isHoverable: {
-        true: `cursor-pointer transition`,
+        true: 'cursor-pointer transition',
       },
       hasNoWidth: {
-        true: `w-full`,
+        true: 'w-full',
       },
     },
     compoundVariants: [
       {
         shadow: 'small',
         isHoverable: true,
-        class: `hover:shadow-neutral-drop-shadow-1xl`,
+        class: 'hover:shadow-neutral-drop-shadow-1xl',
       },
       {
         shadow: 'medium',
         isHoverable: true,
-        class: `hover:shadow-neutral-drop-shadow-2xl`,
+        class: 'hover:shadow-neutral-drop-shadow-2xl',
       },
       {
         shadow: 'large',
         isHoverable: true,
-        class: `hover:shadow-neutral-drop-shadow-3xl`,
+        class: 'hover:shadow-neutral-drop-shadow-3xl',
       },
     ],
   })
@@ -159,7 +162,7 @@ export class LuksoCard extends TailwindStyledElement(style) {
       shadow: this.shadow,
       isHoverable: this.isHoverable,
       hasNoWidth: !this.width,
-      class: `grid grid-rows-[auto,1fr] overflow-hidden`,
+      class: 'grid grid-rows-[auto,1fr] overflow-hidden',
     })
 
     return html`
@@ -189,7 +192,7 @@ export class LuksoCard extends TailwindStyledElement(style) {
       shadow: this.shadow,
       isHoverable: this.isHoverable,
       hasNoWidth: !this.width,
-      class: `grid grid-rows-[1fr,auto]`,
+      class: 'grid grid-rows-[1fr,auto]',
     })
 
     return html`
@@ -203,18 +206,23 @@ export class LuksoCard extends TailwindStyledElement(style) {
       >
         <div
           style=${styleMap({
-            backgroundImage: this.backgroundImageOrGradient(),
+            backgroundImage: this.backgroundGradient(),
           })}
           class=${cn(
-            'min-h-[92px] -mb-6 w-full bg-center bg-cover rounded-t-[inherit] rounded-b-0 relative',
+            'min-h-[92px] -mb-6 w-full rounded-t-[inherit] rounded-b-0 relative',
             this.headerClass
           )}
         >
           ${this.backgroundUrl &&
-          html` <div
-            class="h-full full rounded-t-[inherit] rounded-b-0 bg-neutral-20/10 absolute"
-          ></div>`}
-          <div>
+          html`<div
+              class="rounded-t-[inherit] overflow-hidden absolute inset-0"
+            >
+              <lukso-image src=${this.backgroundUrl}></lukso-image>
+            </div>
+            <div
+              class="h-full full rounded-t-[inherit] rounded-b-0 bg-neutral-20/10 absolute"
+            ></div>`}
+          <div class="relative">
             <slot name="header"></slot>
           </div>
         </div>
@@ -251,7 +259,7 @@ export class LuksoCard extends TailwindStyledElement(style) {
       shadow: this.shadow,
       isHoverable: this.isHoverable,
       hasNoWidth: !this.width,
-      class: `grid grid-rows-[auto,1fr]`,
+      class: 'grid grid-rows-[auto,1fr]',
     })
 
     return html`
@@ -265,7 +273,7 @@ export class LuksoCard extends TailwindStyledElement(style) {
       >
         <div
           style=${styleMap({
-            backgroundImage: this.backgroundImageOrGradient(),
+            backgroundImage: this.backgroundGradient(),
           })}
           class=${cn(
             'min-h-[129px] -mb-6 bg-center bg-cover rounded-t-[inherit] rounded-b-0 relative bg-neutral-100',
@@ -274,8 +282,13 @@ export class LuksoCard extends TailwindStyledElement(style) {
         >
           ${this.backgroundUrl &&
           html`<div
-            class="min-h-full min-w-full rounded-t-[inherit] rounded-b-0 bg-neutral-10/10 absolute"
-          ></div>`}
+              class="rounded-t-[inherit] overflow-hidden absolute inset-0"
+            >
+              <lukso-image src=${this.backgroundUrl}></lukso-image>
+            </div>
+            <div
+              class="min-h-full min-w-full rounded-t-[inherit] rounded-b-0 bg-neutral-10/10 absolute"
+            ></div>`}
           <div>
             <slot name="header"></slot>
           </div>
@@ -318,7 +331,7 @@ export class LuksoCard extends TailwindStyledElement(style) {
       shadow: this.shadow,
       isHoverable: this.isHoverable,
       hasNoWidth: !this.width,
-      class: `h-[240px] flex`,
+      class: 'h-[240px] flex',
     })
 
     return html`
@@ -331,17 +344,17 @@ export class LuksoCard extends TailwindStyledElement(style) {
         class=${cn(cardStyles, this.customClass)}
       >
         <div
-          style=${styleMap({
-            backgroundImage: `url('${this.backgroundUrl}')`,
-          })}
           class="h-full w-full -mb-6 bg-center bg-cover rounded-[inherit] relative"
         >
+          <div class="rounded-[inherit] overflow-hidden absolute inset-0">
+            <lukso-image src=${this.backgroundUrl}></lukso-image>
+          </div>
           ${this.backgroundUrl &&
           html`<div
             class="h-full w-full rounded-[inherit] bg-neutral-20/10 absolute"
           ></div>`}
           <div
-            class="h-full w-full flex flex-col items-center justify-center  absolute"
+            class="h-full w-full flex flex-col items-center justify-center absolute"
           >
             <slot name="content"></slot>
           </div>
@@ -356,7 +369,7 @@ export class LuksoCard extends TailwindStyledElement(style) {
       shadow: this.shadow,
       isHoverable: this.isHoverable,
       hasNoWidth: !this.width,
-      class: `grid grid-rows-[auto,1fr]`,
+      class: 'grid grid-rows-[auto,1fr]',
     })
 
     return html`
@@ -370,14 +383,17 @@ export class LuksoCard extends TailwindStyledElement(style) {
       >
         <div
           style=${styleMap({
-            backgroundImage: this.backgroundImageOrGradient(),
+            backgroundImage: this.backgroundGradient(),
           })}
           class=${cn(
             'min-h-[240px] bg-center bg-cover rounded-t-[inherit] rounded-b-0 relative',
             this.headerClass
           )}
         >
-          <div>
+          <div class="rounded-t-[inherit] overflow-hidden absolute inset-0">
+            <lukso-image src=${this.backgroundUrl}></lukso-image>
+          </div>
+          <div class="relative">
             <slot name="header"></slot>
           </div>
         </div>