+++ src/components/lukso-tooltip/index.ts
@@ -1,10 +1,11 @@
-import { PropertyValues, html } from 'lit'
+import { type PropertyValues, html } from 'lit'
 import { customElement, property, state } from 'lit/decorators.js'
 import tippy from 'tippy.js'
+import { tv } from 'tailwind-variants'
+import { styleMap } from 'lit/directives/style-map.js'
 
 import { TailwindStyledElement } from '@/shared/tailwind-element'
 import style from './style.scss?inline'
-import { customClassMap } from '@/shared/directives'
 
 export type TooltipVariant = 'dark' | 'light' | 'success' | 'danger' | 'white'
 export type TooltipSize = 'medium' | 'large'
@@ -26,6 +27,12 @@ export type TooltipPlacement =
   | 'auto-end'
 export type TooltipTrigger = 'mouseenter' | 'click' | 'manual'
 
+export type TooltipOption = {
+  id?: string
+  value: string
+  text: string
+}
+
 @customElement('lukso-tooltip')
 export class LuksoTooltip extends TailwindStyledElement(style) {
   @property({ type: String })
@@ -41,35 +48,47 @@ export class LuksoTooltip extends TailwindStyledElement(style) {
   trigger: TooltipTrigger = 'mouseenter'
 
   @property({ type: String })
-  text: string = ''
+  text = ''
 
   @property({ type: Number, attribute: 'max-width' })
-  maxWidth: number = 300
+  maxWidth = 300
 
   @property({ type: Boolean })
-  show: boolean = false
+  show = false
 
   @property({ type: String, attribute: 'hide-on-click' })
-  hideOnClick: string = 'true'
+  hideOnClick = 'true'
 
   @property({ type: Boolean, attribute: 'is-clipboard-copy' })
-  isClipboardCopy: boolean = false
+  isClipboardCopy = false
 
   @property({ type: String, attribute: 'copy-text' })
-  copyText: string = ''
+  copyText = ''
 
   @property({ type: String, attribute: 'copy-value' })
-  copyValue: string = ''
+  copyValue = ''
 
   @property({ type: Number })
-  offset: number = 10
+  offset = 10
+
+  @property({ type: String })
+  options = ''
 
   @state()
-  showCopy: boolean = false
+  showCopy = false
 
-  private defaultTooltipStyles = 'bg-neutral-20 p-4 hidden'
+  @state()
+  optionsParsed: TooltipOption[] = []
 
   private tooltipInstance = undefined
+
+  private tooltipStyles = tv({
+    slots: {
+      tooltip: 'hidden',
+      trigger: 'cursor-pointer flex flex-col items-center',
+      options: 'rounded-4 hover:bg-neutral-95',
+    },
+  })
 
   private hideOnClickCheck() {
     if (this.hideOnClick === 'toggle') {
@@ -95,7 +114,7 @@ export class LuksoTooltip extends TailwindStyledElement(style) {
       this.tooltipInstance = undefined
     }
 
-    if (!this.text) {
+    if (!this.text && !this.options) {
       return
     }
 
@@ -134,7 +153,7 @@ export class LuksoTooltip extends TailwindStyledElement(style) {
     // when manually trigger tooltip
     if (changedProperties.has('show') && this.trigger === 'manual') {
       if (this.show) {
-        !!!this.tooltipInstance && this.initTooltip()
+        !this.tooltipInstance && this.initTooltip()
         this.tooltipInstance.show()
       } else {
         this.tooltipInstance?.hide()
@@ -142,19 +161,49 @@ export class LuksoTooltip extends TailwindStyledElement(style) {
       return
     }
 
+    if (changedProperties.has('options') && !!this.options) {
+      try {
+        this.optionsParsed = JSON.parse(this.options) as TooltipOption[]
+      } catch (error: unknown) {
+        console.warn('Could not parse options', error)
+      }
+    }
+
     this.initTooltip()
   }
 
+  disconnectedCallback() {
+    super.disconnectedCallback()
+    this.tooltipInstance?.destroy()
+  }
+
+  private optionsTemplate() {
+    const { options } = this.tooltipStyles()
+
+    // because of the bug in the getting styles properly for options we pass them as style property
+    return html`<ul>
+      ${Object.entries(this.optionsParsed)?.map(
+        option =>
+          html`<li
+            class=${options()}
+            style=${styleMap({
+              padding: '4px 8px',
+              cursor: 'pointer',
+            })}
+            onClick="navigator.clipboard.writeText('${option[1].value}')"
+          >
+            ${option[1].text}
+          </li>`
+      )}
+    </ul>`
+  }
+
   render() {
+    const { tooltip, trigger } = this.tooltipStyles()
+
     return html`
-      <div
-        id="tooltip"
-        role="tooltip"
-        class=${customClassMap({
-          [this.defaultTooltipStyles]: true,
-        })}
-      >
-        ${this.text}
+      <div id="tooltip" role="tooltip" class=${tooltip()}>
+        ${this.options ? this.optionsTemplate() : this.text}
       </div>
       ${this.isClipboardCopy
         ? html`<lukso-tooltip
@@ -167,19 +216,11 @@ export class LuksoTooltip extends TailwindStyledElement(style) {
             ?show=${this.showCopy ? true : undefined}
             text=${this.copyText}
           >
-            <div
-              id="trigger"
-              class="cursor-pointer flex flex-col items-center"
-              @click=${this.handleClick}
-            >
+            <div id="trigger" class=${trigger()} @click=${this.handleClick}>
               <slot></slot>
             </div>
           </lukso-tooltip>`
-        : html`<div
-            id="trigger"
-            class="cursor-pointer flex flex-col items-center"
-            @click=${this.handleClick}
-          >
+        : html`<div id="trigger" class=${trigger()} @click=${this.handleClick}>
             <slot></slot>
           </div>`}
     `